+++ semana20/projeto-lama/src/business/UserBusiness.ts
@@ -5,6 +5,7 @@ import { HashManager } from "../services/HashManager";
 import { Authenticator } from "../services/Authenticator";
 import { BaseError } from "../error/BaseError";
 
+
 export class UserBusiness {
 
     constructor(
