+++ semana20/projeto-lama/src/services/Authenticator.ts
@@ -1,8 +1,9 @@
 import * as jwt from "jsonwebtoken";
 
 export class Authenticator {
-  public generateToken(input: AuthenticationData,
-    expiresIn: string = process.env.ACCESS_TOKEN_EXPIRES_IN!): string {
+
+  public generateToken = (input: AuthenticationData,
+    expiresIn: string = process.env.ACCESS_TOKEN_EXPIRES_IN!): string => {
     const token = jwt.sign(
       {
         id: input.id,
@@ -16,7 +17,7 @@ export class Authenticator {
     return token;
   }
 
-  public getData(token: string): AuthenticationData {
+  public getData = (token: string): AuthenticationData => {
     const payload = jwt.verify(token, process.env.JWT_KEY as string) as any;
     const result = {
       id: payload.id,
