+++ semana20/projeto-lama/src/services/IdGenerator.ts
@@ -2,7 +2,7 @@ import { v4 } from "uuid";
 
 export class IdGenerator{
 
-    generate(): string{
+    public generate(): string{
         return v4();
     }
 }
