--- semana20/projeto-lama/tests/servicesMock/HashManager.ts
@@ -1,12 +1,0 @@
-export class HashManager {
-
-    public async hash(text: string): Promise<string> {
-        return "hash"
-    }
-
-    public async compare(text: string, hash: string): Promise<boolean>{
-        return text === hash;
-    }
-}
-
-export default new HashManager()
