--- semana20/projeto-lama/tests/servicesMock/IdGenerator.ts
@@ -1,8 +1,0 @@
-export class IdGenerator{
-
-    generate(): string{
-        return "id";
-    }
-}
-
-export default new IdGenerator()
