+++ semana20/projeto-lama/tests/servicesMock/hashManagerMock.ts
@@ -1,0 +1,12 @@
+export class HashManagerMock {
+
+    public hash = async (text: string): Promise<string> => {
+        return "hash"
+    }
+
+    public compare = async (text: string, hash: string): Promise<boolean> => {
+        return text === hash;
+    }
+}
+
+export default new HashManagerMock()
