+++ semana20/projeto-lama/tests/servicesMock/idGeneratorMock.ts
@@ -1,0 +1,8 @@
+export class IdGeneratorMock{
+
+    public generate(): string{
+        return "id";
+    }
+}
+
+export default new IdGeneratorMock()
